% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_swarm.R
\name{initialize_swarm}
\alias{initialize_swarm}
\title{Initialize Particle Swarm}
\usage{
initialize_swarm(N, xmin, xmax, chaotic = TRUE, mu = 3.99)
}
\arguments{
\item{N}{Integer. The number of particles in the swarm.}

\item{xmin}{Numeric vector of length \code{d}. The minimum bound for each dimension.}

\item{xmax}{Numeric vector of length \code{d}. The maximum bound for each dimension.}
}
\value{
A list containing two elements:
\item{X}{A matrix of size \verb{N x d} representing the initial positions of the particles.}
\item{V}{A matrix of size \verb{N x d} representing the initial velocities of the particles.}
}
\description{
Initialize Particle Swarm
}
\examples{
par(mfrow = c(2, 1))
plot( initialize_swarm(50, c(2, 2), c(3, 3), mu = 3.78)$X[, 1],type = 'l', ylab="Chaotic" )
plot( initialize_swarm(50, c(2, 2), c(3, 3), chaotic = FALSE)$X[, 1], type = 'l', col = 'red',  ylab="Uniform" )
}
