% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSO.R
\name{PSO}
\alias{PSO}
\title{Basic Particle Swarm Optimization Solver}
\arguments{
\item{obj_function}{the name of the function under which the objective function resides}

\item{n}{the number of particles to use in the PSO algorithm}

\item{d}{the number of dimensions (design variables) of the objective function}

\item{bounds}{the bounds of the design space, as a list}

\item{contour}{optional argument to create a 2-D contour plot fo design space.  **Does not work yet!!!}

\item{parallel}{optional argument to state the number of cores to dedicate for parallel processing.  **Does not work yet.}
}
\value{
a list containing the solution and the position of the design variables at that solution
}
\description{
Finds a solution to a multimodal objective function using basic PSO.
}
\examples{
## Define you objective function by name - Ackley's Path Function
obj_function = function(x) {
   f <<- f + 1
   return((4 - 2.1 * x[1]**2 + (x[1]**4) / 3) * x[1]**2 + x[1] * x[2] + (-4 + 4 * x[2]**2) * x[2]**2)
}

## Define your bounds
bounds = list(c(-2.0, 2.0), c(-2.0, 2.0))

## Call the function
p <- PSO(obj_function, n=15, d=2, bounds=bounds)

}
