% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_personal_bests.R
\name{update_personal_bests}
\alias{update_personal_bests}
\title{Update Matrix of Personal Bests by Fitness}
\usage{
update_personal_bests(X, fitness, P, P_fitness)
}
\arguments{
\item{X}{A matrix of size \verb{N x d} representing the current positions of the particles.}

\item{fitness}{A numeric vector of length \code{N} representing the current fitness values of the particles.}

\item{P}{A matrix of size \verb{N x d} representing the current personal best positions of the particles.}

\item{P_fitness}{A numeric vector of length \code{N} representing the current personal best fitness values.}
}
\value{
A list containing two elements:
\item{P}{An updated matrix of size \verb{N x d} representing the personal best positions of the particles.}
\item{P_fitness}{An updated numeric vector of length \code{N} representing the personal best fitness values.}
}
\description{
Update Matrix of Personal Bests by Fitness
}
\examples{
N <- 30
swarm <- initialize_swarm(N, c(-5, -5), c(5, 5))
fitness <- evaluate_fitness(swarm$X, function(x) sum(x^2))
personal_bests <- initialize_personal_bests(swarm$X, fitness)
updated_personal_bests <- update_personal_bests(swarm$X, fitness, personal_bests$P, personal_bests$P_fitness)
}
\seealso{
\code{\link{initialize_personal_bests}}, \code{\link{evaluate_fitness}}
}
