% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{pso}
\alias{pso}
\title{Particle Swarm Optimizer}
\usage{
pso(fn, S, lb, ub, w = 0.1, max_iter = 1000)
}
\arguments{
\item{fn}{Function to be minimized}

\item{S}{Number of particles in swarm}

\item{lb}{Lower bounds of inputs}

\item{ub}{Upper bounds of inputs}

\item{w}{Weight coefficient}

\item{max_iter}{Maximum number of iterations}
}
\value{
List containing the optimal parameters, optimal function value, and number of iterations
}
\description{
This particle swarm optimizer takes in a function and outputs
the optimal parameter values. The function must be formatted in such
a way that it takes a single vector input for all parameters (i.e.
VEC = c(1,2,3) for x=1, y=2, z=3).
}
